fix(wasm): allow log level to be updated at runtime
Previously, calling `set_logging_level` more than once was causing a
panic at runtime.
With this fix, it’s possible to update the log level at runtime. The
logger is also initialized by the `init` function with a default `WARN`
level, so even if `set_logging_level` is not called, there is a
reasonnable default logger loaded.

Issue: DPS-12563